Refactor successful
