+++ app/src/main/AndroidManifest.xml
@@ -25,7 +25,7 @@
                 android:value=".MainActivity" />
         </activity>
         <activity
-            android:name=".PizzaDetailActivity"
+            android:name=".ui.pizza.PizzaDetailActivity"
             android:parentActivityName=".MainActivity">
         </activity>
     </application>
