+++ app/src/main/java/com/hfad/bitsandpizzas/ui/stores/StoresFragment.java
@@ -1,4 +1,4 @@
-package com.hfad.bitsandpizzas;
+package com.hfad.bitsandpizzas.ui.stores;
 
 
 import android.os.Bundle;
@@ -9,18 +9,21 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
 
+import com.hfad.bitsandpizzas.R;
+
 
 /**
  * A simple {@link Fragment} subclass.
  */
-public class PastaFragment extends ListFragment {
+public class StoresFragment extends ListFragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         ArrayAdapter<String> adapter = new ArrayAdapter<>(
                 inflater.getContext(), android.R.layout.simple_list_item_1,
-                getResources().getStringArray(R.array.pasta));
+                getResources().getStringArray(R.array.stores));
         setListAdapter(adapter);
         return super.onCreateView(inflater, container, savedInstanceState);
     }
+
 }
