+++ app/src/main/res/layout/activity_pizza_detail.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="com.hfad.bitsandpizzas.PizzaDetailActivity"
+    tools:context="com.hfad.bitsandpizzas.ui.pizza.PizzaDetailActivity"
     >
     <include
         layout="@layout/toolbar_main"
