+++ app/src/main/res/layout/fragment_top.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.hfad.bitsandpizzas.TopFragment"
+    tools:context="com.hfad.bitsandpizzas.ui.main.TopFragment"
     >
     <FrameLayout
         android:layout_width="match_parent"
